+++ src/app/components/AtomNetwork/Network.tsx
@@ -1,7 +1,7 @@
 import React, {useEffect, useState} from 'react';
 import * as d3 from 'd3';
-import makeRelationshipLinks from '../utils/makeRelationshipLinks';
-import {filteredSnapshot} from '../../types';
+import makeRelationshipLinks from '../../utils/makeRelationshipLinks';
+import {filteredSnapshot} from '../../../types';
 interface NetworkProps {
   filteredCurSnap: filteredSnapshot;
 }
